requirements.txt
aiogram>=3.0.0
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
aiohttp
apscheduler
python-dotenv
